Refactor scenario reporting for models with more than one executable process
