+++ report/src/main/java/org/camunda/bpm/scenario/report/bpmn/AbstractProcessReport.java
@@ -2,23 +2,14 @@ package org.camunda.bpm.scenario.report.bpmn;
 
 import org.camunda.bpm.engine.ProcessEngine;
 import org.camunda.bpm.engine.ProcessEngines;
-import org.camunda.bpm.engine.history.HistoricActivityInstance;
-import org.camunda.bpm.model.bpmn.BpmnModelInstance;
-import org.camunda.bpm.model.bpmn.instance.Process;
 import org.camunda.bpm.scenario.report.Report;
 
-import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 /**
  * @author Martin Schimak
  */
 abstract class AbstractProcessReport<T> implements Report<T> {
-
-  private static final Map<String, List<HistoricActivityInstance>> coverageActivityInstances = new HashMap<>();
-  private static final Map<String, BpmnModelInstance> coverageBpmnModelInstances = new HashMap<>();
-  private static final Map<String, BpmnModelInstance> scenarioBpmnModelInstances = new HashMap<>();
 
   protected ProcessEngine processEngine;
 
@@ -41,42 +32,6 @@ abstract class AbstractProcessReport<T> implements Report<T> {
         "the constructor with the engine as parameter.";
       throw new IllegalStateException(message);
     }
-  }
-
-  protected BpmnModelInstance getBpmnModelInstanceForScenario(String processDefinitionId) {
-    String processDefinitionKey = processEngine.getRepositoryService().createProcessDefinitionQuery()
-      .processDefinitionId(processDefinitionId).singleResult().getKey();
-    BpmnModelInstance bpmnModelInstance = scenarioBpmnModelInstances.get(processDefinitionKey);
-    if (bpmnModelInstance == null)
-      bpmnModelInstance = processEngine.getRepositoryService().getBpmnModelInstance(processDefinitionId);
-    for (Process process : bpmnModelInstance.getDefinitions().getChildElementsByType(Process.class)) {
-      processDefinitionKey = process.getId();
-      if (process.isExecutable()) {
-        scenarioBpmnModelInstances.putIfAbsent(processDefinitionKey, bpmnModelInstance);
-        coverageBpmnModelInstances.putIfAbsent(processDefinitionKey, bpmnModelInstance);
-      }
-    }
-    return bpmnModelInstance;
-  }
-
-  protected BpmnModelInstance getBpmnModelInstanceForCoverage(String processDefinitionKey) {
-    return coverageBpmnModelInstances.get(processDefinitionKey).clone();
-  }
-
-  protected List<HistoricActivityInstance> findActivityInstancesByProcessInstanceId(String processInstanceId) {
-    List<HistoricActivityInstance> activityInstanceList = processEngine.getHistoryService().createHistoricActivityInstanceQuery()
-      .processInstanceId(processInstanceId).list();
-    if (!activityInstanceList.isEmpty()) {
-      String processDefinitionKey = activityInstanceList.iterator().next().getProcessDefinitionKey();
-      if (coverageActivityInstances.putIfAbsent(processDefinitionKey, activityInstanceList) != null) {
-        coverageActivityInstances.get(processDefinitionKey).addAll(activityInstanceList);
-      }
-    }
-    return activityInstanceList;
-  }
-
-  protected List<HistoricActivityInstance> findActivityInstancesByProcessDefinitionKey(String processDefinitionKey) {
-    return coverageActivityInstances.get(processDefinitionKey);
   }
 
 }
